Reformat tower sub class code and add more memory safety
There were multiple methods that did the same and memory safety could not
be guaranteed since the sub class could override the dispose method from the
abstract class.
Also the play state needed to remove tower projectiles from the world which
is much better suited to be placed in the tower implementation itself.