Rework the zombie subclasses (and in turn rework scoring)
just like the tower subclasses which should make it easier and better to add new
enemies and initialize them.
Through value initialization and final/private addition it should also be much
easier understanding the class since some possible side effects are now gone.
The previous scoring was based on the path the zombie still had to go to the
trailer but this is for now replaced by just the given score value since it gives
hilariously many points for kills that happen at the beginning (like 200 points
instead of 2).